refactor(angular): remove .component from file names
