+++ src/angular/captcha/captcha.module.ts
@@ -1,8 +1,8 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 
+import { SbbCaptcha } from './captcha/captcha';
 import { SbbCaptchaLoaderService } from './captcha/captcha-loader.service';
-import { SbbCaptcha } from './captcha/captcha.component';
 import { SbbWindowRef } from './captcha/windowref.service';
 
 @NgModule({
