+++ src/angular/captcha/captcha/captcha.ts
