+++ src/angular/captcha/public-api.ts
@@ -2,4 +2,4 @@ export * from './captcha.module';
 export * from './captcha/captcha-settings';
 export * from './captcha/windowref.service';
 export * from './captcha/captcha-loader.service';
-export * from './captcha/captcha.component';
+export * from './captcha/captcha';
