+++ src/angular/checkbox/checkbox.html
