+++ src/angular/checkbox/checkbox.module.ts
@@ -1,8 +1,8 @@
 import { ObserversModule } from '@angular/cdk/observers';
 import { NgModule } from '@angular/core';
 
+import { SbbCheckbox } from './checkbox';
 import { SbbCheckboxRequiredValidator } from './checkbox-required-validator';
-import { SbbCheckbox } from './checkbox.component';
 
 @NgModule({
   imports: [ObserversModule],
