+++ src/angular/checkbox/checkbox.spec.ts
