+++ src/angular/checkbox/public-api.ts
@@ -1,4 +1,4 @@
 export * from './checkbox-config';
 export * from './checkbox-required-validator';
-export * from './checkbox.component';
+export * from './checkbox';
 export * from './checkbox.module';
