+++ src/angular/loading/BUILD.bazel
@@ -15,7 +15,7 @@ ng_module(
         ["**/*.ts"],
         exclude = ["**/*.spec.ts"],
     ),
-    assets = [":loading/loading.component.css"] + glob(["**/*.html"]),
+    assets = [":loading/loading.css"] + glob(["**/*.html"]),
     module_name = "@sbb-esta/angular/loading",
     deps = [
         "@npm//@angular/common",
@@ -24,8 +24,8 @@ ng_module(
 )
 
 sass_binary(
-    name = "loading_loading_component_scss",
-    src = "loading/loading.component.scss",
+    name = "loading_loading_scss",
+    src = "loading/loading.scss",
     deps = [
         "//src/angular/styles:common_scss_lib",
     ],
