+++ src/angular/loading/loading.module.ts
@@ -1,7 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 
-import { SbbLoading } from './loading/loading.component';
+import { SbbLoading } from './loading/loading';
 
 @NgModule({
   imports: [CommonModule],
