+++ src/angular/loading/loading/loading.html
