+++ src/angular/loading/loading/loading.scss
