+++ src/angular/loading/loading/loading.ts
@@ -2,8 +2,8 @@ import { ChangeDetectionStrategy, Component, Input, ViewEncapsulation } from '@a
 
 @Component({
   selector: 'sbb-loading',
-  templateUrl: './loading.component.html',
-  styleUrls: ['./loading.component.css'],
+  templateUrl: './loading.html',
+  styleUrls: ['./loading.css'],
   changeDetection: ChangeDetectionStrategy.OnPush,
   encapsulation: ViewEncapsulation.None,
   host: {
