+++ src/angular/radio-button/public-api.ts
@@ -1,2 +1,2 @@
 export * from './radio-button.module';
-export * from './radio-button.component';
+export * from './radio-button';
