+++ src/angular/textarea/BUILD.bazel
@@ -15,7 +15,7 @@ ng_module(
         ["**/*.ts"],
         exclude = ["**/*.spec.ts"],
     ),
-    assets = [":textarea/textarea.component.css"] + glob(["**/*.html"]),
+    assets = [":textarea/textarea.css"] + glob(["**/*.html"]),
     module_name = "@sbb-esta/angular/textarea",
     deps = [
         "//src/angular/core",
@@ -29,8 +29,8 @@ ng_module(
 )
 
 sass_binary(
-    name = "textarea_textarea_component_scss",
-    src = "textarea/textarea.component.scss",
+    name = "textarea_textarea_scss",
+    src = "textarea/textarea.scss",
     deps = [
         "//src/angular/styles:common_scss_lib",
     ],
