+++ src/angular/textarea/public-api.ts
@@ -1,2 +1,2 @@
 export * from './textarea.module';
-export * from './textarea/textarea.component';
+export * from './textarea/textarea';
