+++ src/angular/textarea/textarea.module.ts
@@ -2,7 +2,7 @@ import { TextFieldModule } from '@angular/cdk/text-field';
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 
-import { SbbTextarea } from './textarea/textarea.component';
+import { SbbTextarea } from './textarea/textarea';
 
 @NgModule({
   imports: [CommonModule, TextFieldModule],
