+++ src/angular/textarea/textarea/textarea.scss
